refactor: extract onboarding components to fix page export error
